pretty-up artfix (#22)
* pretty-up and add no-op to artfix

+++ src/commands/artfix.rs
@@ -1,33 +1,45 @@
 use crate::utils::config::MAX_ARTWORK_SIZE;
 use crate::utils::dir::expand_dir_list;
+use crate::utils::types::GlobalOpts;
 use image::imageops::FilterType::Lanczos3;
 use image::GenericImageView;
 use image::ImageReader;
+use indicatif::ProgressBar;
 use std::fs;
 use std::os::unix::fs::symlink;
 use std::path::{Path, PathBuf};
 
-pub fn run(dirlist: &[String], recurse: bool, linkdir: String) -> anyhow::Result<()> {
+pub fn run(
+    dirlist: &[String],
+    recurse: bool,
+    linkdir: String,
+    opts: &GlobalOpts,
+) -> anyhow::Result<()> {
     let linkdir = PathBuf::from(linkdir);
     let dirlist: Vec<PathBuf> = dirlist.iter().map(PathBuf::from).collect();
 
-    for dir in expand_dir_list(&dirlist, recurse) {
-        check_artwork(&dir, &linkdir)?;
+    let dirs = expand_dir_list(&dirlist, recurse);
+    let bar = ProgressBar::new(dirs.len() as u64);
+
+    for dir in dirs {
+        bar.inc(1);
+        check_artwork(&dir, &linkdir, opts)?;
     }
+    bar.finish();
     Ok(())
 }
 
-fn check_artwork(dir: &Path, linkdir: &Path) -> anyhow::Result<bool> {
+fn check_artwork(dir: &Path, linkdir: &Path, opts: &GlobalOpts) -> anyhow::Result<bool> {
     let expected_artwork = dir.join("front.jpg");
 
     let mut changes: Vec<bool> = Vec::new();
 
     if !expected_artwork.exists() {
-        changes.push(rename(dir, &expected_artwork)?);
+        changes.push(rename(dir, &expected_artwork, opts)?);
     }
 
     if expected_artwork.exists() {
-        changes.push(resize_or_link(&expected_artwork, linkdir)?);
+        changes.push(resize_or_link(&expected_artwork, linkdir, opts)?);
     }
 
     if changes.iter().any(|c| *c) {
@@ -37,14 +49,16 @@ fn check_artwork(dir: &Path, linkdir: &Path) -> anyhow::Result<bool> {
     }
 }
 
-fn rename(dir: &Path, front: &Path) -> anyhow::Result<bool> {
+fn rename(dir: &Path, front: &Path, opts: &GlobalOpts) -> anyhow::Result<bool> {
     let mut ret = false;
 
     // Yes, this will flatten multiple files into one. That's fine.
     for file in jpgs_in(dir)? {
         if file != front {
-            println!("moving {} to {}", file.display(), front.display());
-            fs::rename(file, front)?;
+            println!("Rename: {} -> front.jpg", file.display());
+            if !opts.noop {
+                fs::rename(file, front)?;
+            }
             ret = true;
         }
     }
@@ -54,12 +68,12 @@ fn rename(dir: &Path, front: &Path) -> anyhow::Result<bool> {
 
 // It's not our job to flag up problems, only to fix what we can. Lintdir will point out what
 // we can't fix.
-fn resize_or_link(file: &Path, linkdir: &Path) -> anyhow::Result<bool> {
+fn resize_or_link(file: &Path, linkdir: &Path, opts: &GlobalOpts) -> anyhow::Result<bool> {
     let img = ImageReader::open(file)?.decode()?;
     let (x, y) = img.dimensions();
 
     if x != y {
-        return symlink_art(file, linkdir);
+        return symlink_art(file, linkdir, opts);
     }
 
     if x <= MAX_ARTWORK_SIZE {
@@ -67,12 +81,16 @@ fn resize_or_link(file: &Path, linkdir: &Path) -> anyhow::Result<bool> {
     }
 
     println!(
-        "Rescaling {} to {s}x{s}",
+        "Resize: {} -> {s}x{s}",
         file.display(),
         s = MAX_ARTWORK_SIZE
     );
-    let new_img = img.resize_exact(MAX_ARTWORK_SIZE, MAX_ARTWORK_SIZE, Lanczos3);
-    new_img.save(file)?;
+
+    if !opts.noop {
+        let new_img = img.resize_exact(MAX_ARTWORK_SIZE, MAX_ARTWORK_SIZE, Lanczos3);
+        new_img.save(file)?;
+    }
+
     Ok(true)
 }
 
@@ -83,20 +101,29 @@ fn target_filename(file: &Path) -> String {
         .into()
 }
 
-fn symlink_art(file: &Path, linkdir: &Path) -> anyhow::Result<bool> {
+fn symlink_art(file: &Path, linkdir: &Path, opts: &GlobalOpts) -> anyhow::Result<bool> {
     if !linkdir.exists() {
         println!("Creating artfix dir: {}", linkdir.display());
-        fs::create_dir_all(linkdir)?;
+        if !opts.noop {
+            fs::create_dir_all(linkdir)?;
+        }
     }
 
     let target = linkdir.join(target_filename(file));
 
-    if target.exists() {
+    if target.exists() && !opts.noop {
         fs::remove_file(&target)?;
     }
 
-    println!("Linking {} -> {}", file.display(), target.display());
-    symlink(file, &target)?;
+    println!(
+        "Symlink: {} -> {}",
+        file.file_name().unwrap().to_string_lossy(),
+        target.display()
+    );
+
+    if !opts.noop {
+        symlink(file, &target)?;
+    }
     Ok(true)
 }
 
@@ -117,14 +144,15 @@ fn jpgs_in(dir: &Path) -> anyhow::Result<Vec<PathBuf>> {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::utils::spec_helper::fixture;
+    use crate::utils::spec_helper::{defopts, fixture};
     use assert_fs::prelude::*;
 
     #[test]
     fn test_resize_no_action() {
         assert!(!resize_or_link(
             &fixture("commands/artfix/tester.good_art/front.jpg"),
-            &PathBuf::from("/tmp")
+            &PathBuf::from("/tmp"),
+            &defopts()
         )
         .unwrap());
     }
@@ -146,7 +174,7 @@ mod test {
         assert_eq!(x, 900);
         assert_eq!(y, 900);
 
-        assert!(resize_or_link(&file_under_test, &linkdir).unwrap());
+        assert!(resize_or_link(&file_under_test, &linkdir, &defopts()).unwrap());
 
         let after = ImageReader::open(&file_under_test)
             .unwrap()
@@ -162,7 +190,7 @@ mod test {
         let source_file = fixture("commands/artfix/tester.not_square/front.jpg");
         let target_dir = assert_fs::TempDir::new().unwrap();
         let expected_file = target_dir.join(target_filename(&source_file));
-        assert!(resize_or_link(&source_file, &target_dir).unwrap());
+        assert!(resize_or_link(&source_file, &target_dir, &defopts()).unwrap());
 
         assert!(expected_file.exists());
         assert!(expected_file.is_symlink());
@@ -180,10 +208,10 @@ mod test {
         assert!(dir_under_test.join("cover.jpg").exists());
         assert!(!expected_artwork.exists());
 
-        assert!(rename(&dir_under_test, &expected_artwork).unwrap());
+        assert!(rename(&dir_under_test, &expected_artwork, &defopts()).unwrap());
         assert!(!dir_under_test.join("cover.jpg").exists());
         assert!(expected_artwork.exists());
 
-        assert!(!rename(&dir_under_test, &expected_artwork).unwrap());
+        assert!(!rename(&dir_under_test, &expected_artwork, &defopts()).unwrap());
     }
 }
